[SBS-4][SBS-20] ui: list review
