+++ app/components/index.ts
@@ -12,5 +12,6 @@ export * from "./Toggle";
 export * from "./EmptyState";
 export * from "./Dropdown";
 export * from "./RadioGroup";
+export * from "./Rating";
 export * from "./ImagePicker";
 export * from "./CreateProduct/SelectImageGroup";
