+++ app/screens/DemoHomePageScreen/DemoHomePageScreen.tsx
@@ -1,16 +1,14 @@
 import React, { FC, ReactElement, useEffect, useState } from "react";
-import { Image, TextStyle, View, ViewStyle } from "react-native";
-import { Screen, Text, TextField} from "../../components";
+import { Image, ImageStyle, TextStyle, View, ViewStyle } from "react-native";
+import { Screen, Text, TextField } from "../../components";
 import { DemoTabScreenProps } from "../../navigators/DemoNavigator";
 import { colors, spacing } from "../../theme";
 import { ScrollView } from "react-native-gesture-handler";
 import { api } from "../../services/api";
 import type { Category } from "app/types";
 import SlideShow from "app/components/SlideShow";
+import { Ionicons } from "@expo/vector-icons";
 
-const iconSearch = require("../../../assets/icons/search.png");
-const iconRight = require("../../../assets/icons/arrowRight.png");
-const iconClothes = require("../../../assets/icons/clothes.png");
 const imgProduct = require("../../../assets/icons/product.jpg");
 export interface Demo {
   name: string;
@@ -18,10 +16,8 @@ export interface Demo {
   data: ReactElement[];
 }
 
-
 export const DemoHomePageScreen: FC<DemoTabScreenProps<"DemoHomePage">> =
   function DemoHomePageScreen(_props) {
-
     const images = [
       "https://via.placeholder.com/600/92c952",
       "https://via.placeholder.com/600/771796",
@@ -33,7 +29,6 @@ export const DemoHomePageScreen: FC<DemoTabScreenProps<"DemoHomePage">> =
     const [categories, setCategories] = useState<Category[]>([]);
 
     useEffect(() => {
-
       const fetchCategories = async () => {
         try {
           const response = await api.getCategories();
@@ -48,136 +43,125 @@ export const DemoHomePageScreen: FC<DemoTabScreenProps<"DemoHomePage">> =
     }, []);
 
     return (
-        <Screen preset="scroll" safeAreaEdges={["top"]} contentContainerStyle={$screenContainer}>
-              <View style={$search}>
-                <Image source={iconSearch} style={$iconSearch} />
-                <TextField
-                  style={$inputSearch}
-                  placeholderTx="demoHomePageScreen.placeholderSearch"
-                />
-              </View>
+      <Screen preset="scroll" safeAreaEdges={["top"]} contentContainerStyle={$screenContainer}>
+        <View style={$search}>
+          <Ionicons style={$iconSearch} name="search-outline" size={20} color="black" />
+          <TextField style={$inputSearch} placeholderTx="demoHomePageScreen.placeholderSearch" />
+        </View>
 
-              <View style={$listCategory}>
-                <View style={$categoryHead}>
-                    <Text
-                      tx="demoHomePageScreen.categoryList"
-                      weight="bold"
-                    />
-                  <View style={$categoryNav}>
-                    <Text
-                      tx="demoHomePageScreen.viewAll"
-                    />
-                    <Image source={iconRight} style={$iconSearch} />
+        <View style={$listCategory}>
+          <View style={$categoryHead}>
+            <Text tx="demoHomePageScreen.categoryList" weight="bold" />
+            <View style={$categoryNav}>
+              <Text tx="demoHomePageScreen.viewAll" />
+              <Ionicons name="chevron-forward-outline" size={20} color="black" />
+            </View>
+          </View>
+          <View>
+            <ScrollView
+              style={$categoryBody}
+              horizontal={true}
+              showsHorizontalScrollIndicator={false}
+            >
+              {categories.length > 0 &&
+                categories.map((category, index) => (
+                  <View key={index} style={$categoryList}>
+                    <View style={$categoryItem}>
+                      <Ionicons name="body-outline" size={20} color="black" />
+                    </View>
+                    <Text style={$categoryText}>{category.name}</Text>
                   </View>
+                ))}
+            </ScrollView>
+          </View>
+        </View>
+
+        <SlideShow images={images} />
+
+        <View style={$listProduct}>
+          <View style={$categoryHead}>
+            <Text tx="demoHomePageScreen.recommended" weight="bold" />
+            <View style={$categoryNav}>
+              <Text tx="demoHomePageScreen.viewAll" />
+              <Ionicons name="chevron-forward-outline" size={20} color="black" />
+            </View>
+          </View>
+
+          <View style={$listProductBody}>
+            <View style={$productItem}>
+              <Image source={imgProduct} style={$imgProduct} />
+              <View style={$productItemContain}>
+                <Text style={$productName}>Adidas white sneakers for men</Text>
+                <View style={$productItemContain}>
+                  <Text weight="bold" style={$productPrice}>
+                    $68
+                  </Text>
+                  <Text style={$productSale}>$134</Text>
+                  <Text style={$productDiscount}>50% OFF</Text>
                 </View>
-                  <View >
-                  <ScrollView
-                    style={$categoryBody} horizontal={true} showsHorizontalScrollIndicator={false}>
-                    { categories.length > 0 &&
-                      categories.map((category, index) => (
-                        <View key={index} style={$categoryList}>
-                          <View style={$categoryItem}>
-                            <Image source={iconClothes} style={$iconSearch} />
-                          </View>
-                          <Text style={$categoryText}>{category.name}</Text>
-                          <Text>{category.name}</Text>
-                      </View>
-                      ))
-                    }
-                  </ScrollView>
-                  </View>
-              </View>
-
-              <SlideShow images={images} />
-
-              <View style={$listProduct}>
-                <View style={$categoryHead}>
-                  <Text
-                    tx="demoHomePageScreen.recommended"
-                    weight="bold"
-                  />
-                  <View style={$categoryNav}>
-                    <Text
-                      tx="demoHomePageScreen.viewAll"
-                    />
-                    <Image source={iconRight} style={$iconSearch} />
-                  </View>
-                </View>
-
-                <View style={$listProductBody}>
-                  <View style={$productItem}>
-                    <Image source={imgProduct} style={$imgProduct} />
-                    <View style={$productItemContain}>
-                      <Text style={$productName}>Adidas white sneakers for men</Text>
-                      <View style={$productItemContain}>
-                        <Text weight="bold" style={$productPrice}>$68</Text>
-                        <Text style={$productSale}>$134</Text>
-                        <Text style={$productDiscount}>50% OFF</Text>
-                      </View>
-                      <View style={$productRatingContain}>
-                        <Image source={imgProduct} style={$iconstar} />
-                        <Text style={$productRating}>4.8</Text>
-                        <Text style={$productComment}>(545)</Text>
-                      </View>
-                    </View>
-                  </View>
-
-                  <View style={$productItem}>
-                    <Image source={imgProduct} style={$imgProduct} />
-                    <View style={$productItemContain}>
-                      <Text style={$productName}>Adidas white sneakers for men</Text>
-                      <View style={$productItemContain}>
-                        <Text style={$productPrice}>$68</Text>
-                        <Text style={$productSale}>$134</Text>
-                        <Text style={$productDiscount}>50% OFF</Text>
-                      </View>
-                      <View style={$productRatingContain}>
-                        <Image source={imgProduct} style={$iconstar} />
-                        <Text style={$productRating}>4.8</Text>
-                        <Text style={$productComment}>(545)</Text>
-                      </View>
-                    </View>
-                  </View>
-
-                  <View style={$productItem}>
-                    <Image source={imgProduct} style={$imgProduct} />
-                    <View style={$productItemContain}>
-                      <Text style={$productName}>Adidas white sneakers for men</Text>
-                      <View style={$productItemContain}>
-                        <Text style={$productPrice}>$68</Text>
-                        <Text style={$productSale}>$134</Text>
-                        <Text style={$productDiscount}>50% OFF</Text>
-                      </View>
-                      <View style={$productRatingContain}>
-                        <Image source={imgProduct} style={$iconstar} />
-                        <Text style={$productRating}>4.8</Text>
-                        <Text style={$productComment}>(545)</Text>
-                      </View>
-                    </View>
-                  </View>
-
-                  <View style={$productItem}>
-                    <Image source={imgProduct} style={$imgProduct} />
-                    <View style={$productItemContain}>
-                      <Text style={$productName}>Adidas white sneakers for men</Text>
-                      <View style={$productItemContain}>
-                        <Text style={$productPrice}>$68</Text>
-                        <Text style={$productSale}>$134</Text>
-                        <Text style={$productDiscount}>50% OFF</Text>
-                      </View>
-                      <View style={$productRatingContain}>
-                        <Image source={imgProduct} style={$iconstar} />
-                        <Text style={$productRating}>4.8</Text>
-                        <Text style={$productComment}>(545)</Text>
-                      </View>
-                    </View>
-                  </View>
+                <View style={$productRatingContain}>
+                  <Ionicons name="star" size={20} color="yellow" />
+                  <Text style={$productRating}>4.8</Text>
+                  <Text style={$productComment}>(545)</Text>
                 </View>
               </View>
-        </Screen>
+            </View>
+
+            <View style={$productItem}>
+              <Image source={imgProduct} style={$imgProduct} />
+              <View style={$productItemContain}>
+                <Text style={$productName}>Adidas white sneakers for men</Text>
+                <View style={$productItemContain}>
+                  <Text style={$productPrice}>$68</Text>
+                  <Text style={$productSale}>$134</Text>
+                  <Text style={$productDiscount}>50% OFF</Text>
+                </View>
+                <View style={$productRatingContain}>
+                  <Ionicons name="star" size={20} color="yellow" />
+                  <Text style={$productRating}>4.8</Text>
+                  <Text style={$productComment}>(545)</Text>
+                </View>
+              </View>
+            </View>
+
+            <View style={$productItem}>
+              <Image source={imgProduct} style={$imgProduct} />
+              <View style={$productItemContain}>
+                <Text style={$productName}>Adidas white sneakers for men</Text>
+                <View style={$productItemContain}>
+                  <Text style={$productPrice}>$68</Text>
+                  <Text style={$productSale}>$134</Text>
+                  <Text style={$productDiscount}>50% OFF</Text>
+                </View>
+                <View style={$productRatingContain}>
+                  <Ionicons name="star" size={20} color="yellow" />
+                  <Text style={$productRating}>4.8</Text>
+                  <Text style={$productComment}>(545)</Text>
+                </View>
+              </View>
+            </View>
+
+            <View style={$productItem}>
+              <Image source={imgProduct} style={$imgProduct} />
+              <View style={$productItemContain}>
+                <Text style={$productName}>Adidas white sneakers for men</Text>
+                <View style={$productItemContain}>
+                  <Text style={$productPrice}>$68</Text>
+                  <Text style={$productSale}>$134</Text>
+                  <Text style={$productDiscount}>50% OFF</Text>
+                </View>
+                <View style={$productRatingContain}>
+                  <Ionicons name="star" size={20} color="yellow" />
+                  <Text style={$productRating}>4.8</Text>
+                  <Text style={$productComment}>(545)</Text>
+                </View>
+              </View>
+            </View>
+          </View>
+        </View>
+      </Screen>
     );
   };
-
 
 const $screenContainer: ViewStyle = {
   backgroundColor: colors.palette.neutral100,
@@ -194,12 +178,8 @@ const $search: ViewStyle = {
 };
 
 const $iconSearch: ViewStyle = {
-  height: 22,
-  width: 22,
   margin: spacing.sm,
 };
-
-
 
 const $inputSearch: TextStyle = {
   width: 350,
@@ -213,14 +193,14 @@ const $categoryHead: ViewStyle = {
   flexDirection: "row",
   justifyContent: "space-between",
   alignItems: "center",
+  marginBottom: spacing.md,
 };
 
 const $categoryNav: ViewStyle = {
   flexDirection: "row",
-   alignItems: "center",
-    gap: -4,
+  alignItems: "center",
+  gap: -4,
 };
-
 
 const $categoryBody: ViewStyle = {
   flexDirection: "row",
@@ -229,7 +209,6 @@ const $categoryBody: ViewStyle = {
 
 const $categoryList: ViewStyle = {
   paddingHorizontal: 10,
-
 };
 
 const $categoryItem: ViewStyle = {
@@ -248,7 +227,7 @@ const $categoryText: TextStyle = {
   maxWidth: 120,
   textAlign: "center",
 };
-const $imgProduct: ViewStyle = {
+const $imgProduct: ImageStyle = {
   width: "100%",
   height: 140,
   borderRadius: spacing.xs,
@@ -259,7 +238,7 @@ const $listProduct: ViewStyle = {
 };
 
 const $listProductBody: ViewStyle = {
-  flexDirection:"row",
+  flexDirection: "row",
   width: "100%",
   gap: 10,
   flexWrap: "wrap",
@@ -271,17 +250,12 @@ const $productItem: ViewStyle = {
 
 const $productRatingContain: ViewStyle = {
   flexDirection: "row",
-   gap: 6,
-    alignItems: "center",
+  gap: 6,
+  alignItems: "center",
 };
 const $productItemContain: ViewStyle = {
   flexDirection: "column",
-   gap: 6,
-};
-
-const $iconstar: ViewStyle = {
-  width: 18,
-  height: 18,
+  gap: 6,
 };
 
 const $productName: TextStyle = {
@@ -298,7 +272,6 @@ const $productPrice: TextStyle = {
 const $productSale: TextStyle = {
   fontSize: 14,
   color: "#ccc",
-
 };
 
 const $productDiscount: TextStyle = {
@@ -311,6 +284,4 @@ const $productRating: ViewStyle = {
   marginTop: 6,
 };
 
-const $productComment: ViewStyle = {
-
-};
+const $productComment: ViewStyle = {};
