+++ app/screens/ProductDetailScreen/ProductDetailScreen.tsx
@@ -3,7 +3,7 @@ import { ScrollView, View, TouchableOpacity, TextStyle, ViewStyle } from "react-
 import { Ionicons } from "@expo/vector-icons";
 
 import SlideShow from "app/components/SlideShow";
-import ListRating from "app/components/ListRating";
+import Rating from "app/components/Rating";
 import { Text } from "app/components";
 import { colors } from "app/theme";
 import CustomHeader from "app/components/CustomHeader";
@@ -92,7 +92,7 @@ const ProductDetailScreen = () => {
             ))}
           </View>
           <View>
-            <ListRating images={images} />
+            <Rating images={images} />
           </View>
         </View>
       </ScrollView>
@@ -114,9 +114,7 @@ const ProductDetailScreen = () => {
           />
         </TouchableOpacity>
       </View>
-
     </View>
-
   );
 };
